+++ inc/commondbtm.class.php
@@ -142,6 +142,15 @@ class CommonDBTM extends CommonGLPI {
    protected $usenotepad = false;
 
    /**
+    * Flag to determine whether or not queued notifications should be deduplicated.
+    * Deduplication is done when a new notification is raised.
+    * Any existing notification for same object, event and recipient is dropped to be replaced by the new one.
+    *
+    * @var boolean
+    */
+   public $deduplicate_queued_notifications = true;
+
+   /**
     * Flag to determine whether or not notification queu should be flushed immediately when an
     * action is performed on item.
     *
@@ -1194,6 +1203,7 @@ class CommonDBTM extends CommonGLPI {
                   $this->clearSavedInput();
                }
                if ($this->notificationqueueonaction) {
+                  Toolbox::deprecated('$notificationqueueonaction property usage is deprecated');
                   QueuedNotification::forceSendFor($this->getType(), $this->fields['id']);
                }
                return $this->fields['id'];
@@ -1573,6 +1583,7 @@ class CommonDBTM extends CommonGLPI {
             }
 
             if ($this->notificationqueueonaction) {
+               Toolbox::deprecated('$notificationqueueonaction property usage is deprecated');
                QueuedNotification::forceSendFor($this->getType(), $this->fields['id']);
             }
 
@@ -1874,6 +1885,7 @@ class CommonDBTM extends CommonGLPI {
                Plugin::doHook("item_delete", $this);
             }
             if ($this->notificationqueueonaction) {
+               Toolbox::deprecated('$notificationqueueonaction property usage is deprecated');
                QueuedNotification::forceSendFor($this->getType(), $this->fields['id']);
             }
 
@@ -2045,6 +2057,7 @@ class CommonDBTM extends CommonGLPI {
          }
          Plugin::doHook("item_restore", $this);
          if ($this->notificationqueueonaction) {
+            Toolbox::deprecated('$notificationqueueonaction property usage is deprecated');
             QueuedNotification::forceSendFor($this->getType(), $this->fields['id']);
          }
          return true;